Crate optra

An engine for keeping remote siles synchronized.

The engine is based on the Admissabilty Based Sequence Transformation algorithm (doi: 10.1109/TPDS.2010.64) The idea is that each site that wants to have a file synchronized will have an instance of Engine running. Any changes that are made to teh file should be run through the engine using either process_diffs() or process_transaction() prior to being broadcast. Any changes that are made at a remote site should be run through the engine using integrate_remote() prior to being applied to the file.

This crate generally works well with rdiff, but can work with any system that generates difference operations that are limited to insert and delete.

Examples

Assuming you have a method read_transaction for reading transactions from remote sites, the process for integrating remote changes onto a local file go like this:

 let mut remote_sequence = read_transaction();
 engine.integrate_remote(&mut remote_sequence);
 remote_sequence.apply("local_file").unwrap();

On the other hand, if you have a Diff generated by detecting differences to a local file, you can process the diff and then send it out (assuming you have a method called send_transaction())

 let diffs = file_hashes.diff_and_update(File::open("local_file").unwrap()).unwrap();
 let transaction = engine.process_diffs(diffs);
 send_transaction(transaction);

Structs

DeleteOperation
Represents an operation which removes data from a file

Engine
Process file change operations in such a way that they can be synchronized across sites

InsertOperation
Represents an operation which inserts data into a file

OTError
Represents an error in attempting to synchronize remote operations

State
Represents the state of a document. Essentially a timestamp and a site id.

TransactionSequence
Represents a sequence of transactions that can be performed on a file.

Enums

ErrorKind
Represents the kind of error we encountered synchronizing operations

Traits

Operation
An operation that will make a change to a file.