Struct optra::State

pub struct State {
    // some fields omitted
}

Represents the state of a document. Essentially a timestamp and a site id.

The state has two timestamps, the remtoe timestamp (the stamp given to it by the site that originated it) and its local timestamp (which is the timestamp this site gave to it)

Methods

impl State

fn new(site_id: u32, local_time: u32, remote_time: u32) -> State

Create a new state at the given site and give it the corresponding timestamps

fn set_time(&mut self, time_stamp: u32)

Sets the local time to the given time stamp

fn get_time(&self) -> u32

Gets the local timestamp for this state

fn matches(&self, other_state: &State) -> bool

Checks if this state (stored locally) happened at the same time as another state (from a remote site).

fn happened_after(&self, other_time: u32, other_id: u32) -> bool

Checks if this state happened after a certain timestamp on a differenent site

fn get_site_id(&self) -> u32

Gets the site id of the origin of this state