Trait optra::Operation

pub trait Operation: Debug + Ord + Clone {
    fn get_state(&self) -> &State;
    fn get_state_mut(&mut self) -> &mut State;
    fn get_position(&self) -> u64;
    fn get_increment(&self) -> i64;
}

An operation that will make a change to a file.

Required Methods

fn get_state(&self) -> &State

Gets the State that this operation was performed on

fn get_state_mut(&mut self) -> &mut State

Gets the State that this operation was performed on mutably.

fn get_position(&self) -> u64

Gets the position this operation will be perfomed at

fn get_increment(&self) -> i64

Gets the size change this operation will perform. For insert operations, it's the length of the data they will insert. For delete operations, it's the length of the data they will delete