Struct optra::TransactionSequence

pub struct TransactionSequence {
    // some fields omitted
}

Represents a sequence of transactions that can be performed on a file.

The operations are stored in the order they occur in the file, and each operation assumes the previous operations have taken place. Furthermore, the insertion operations are assumed to take place prior to the deletion operations.

Methods

impl TransactionSequence

fn new(starting_state: Option<State>, site: u32, inserts: LinkedList<InsertOperation>, deletes: LinkedList<DeleteOperation>) -> TransactionSequence

Construct a new TransactionSequence from the given operations and metadata The starting_state is the state of the file (as represented by a time_stamp) at the time the sequence was created, or None if the file was newly created. The site is the location this TransactionSequence came from.

fn apply<P: AsRef<Path> + Clone>(&self, path: P) -> Result<()>

Apply the operations in this sequence to a file. This should not be called until after the sequence has been integrated via Engine::integrate_remote